Mise a jour finalisation v1

+++ src/commentaires/commentaires.service.ts
@@ -501,10 +501,12 @@ export class CommentairesService {
 
         if ('keys' in store && typeof store.keys === 'function') {
           const keys = await store.keys('commentaires:*');
-          if (keys.length > 0) {
-            if ('delete' in store && typeof store.delete === 'function') {
-              await Promise.all(keys.map((key) => store.delete(key)));
-            }
+          if (
+            keys.length > 0 &&
+            'delete' in store &&
+            typeof store.delete === 'function'
+          ) {
+            await Promise.all(keys.map((key) => store.delete(key)));
           }
         }
       }
